+++ sd-editor/src/main/java/net/sevenscales/editor/diagram/drag/AnchorMoveHandler.java
@@ -5,6 +5,6 @@ import net.sevenscales.editor.diagram.drag.AnchorElement;
 
 
 public interface AnchorMoveHandler {
-  void moving(AnchorElement anchorElement, int dx, int dy, long dispachSequence);
+  void moving(AnchorElement anchorElement, int dx, int dy, int dispachSequence);
   // Relationship2 connection();
 }
